requirements.txt
discord.py==2.3.2       # Version for the latest stable Discord API support
gTTS==2.3.2             # Google Text-to-Speech for TTS functionality
ffmpeg-python==0.2.0    # Python bindings for FFMPEG (optional, used for audio processing)
